Fix #7 Filter out others' cards when analyzing own hand
If all the copies of a given card has been played, discarded, or
drawn by other players, it should not be possible for said card to be
in our own hand.

Similarly, when the last copy of a card has been confirmed to be in
our own hand, the other cards that are in our hand should reuse this
information to perform its own filtering analysis.

Signed-off-by: Remy Suen <remy.suen@gmail.com>