Fix #13 Disable and preselect cards when hinting
If a blue hint has been received and a card is already known to be
blue, that card should be preselected on the hinting page as well as
disabled as it shouldn't be possible to undo it and mark it as not
being blue. Similarly, any card that is known to not be blue should
be unselected and also disabled since it should not be possible to
flag a card that is known to not be blue as blue.

Signed-off-by: Remy Suen <remy.suen@gmail.com>